requirements.txt
alembic==1.3.2
APScheduler==3.6.3
arrow==0.15.5
Babel==2.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloud-init==19.4
command-not-found==0.3
configobj==5.0.6
croniter==0.3.31
cryptography==2.8
dnspython==1.16.0
dominate==2.4.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-paginate==0.5.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.8
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==1.0.3
pandas==0.24.2
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.19
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyodbc==4.0.28
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.8.1
python-debian==0.1.27
python-dotenv==0.10.3
python-editor==1.0.4
python-systemd==231
pytz==2019.3
PyYAML==5.1
redis==3.4.1
requests==2.22.0
rq==1.2.2
rq-scheduler==0.9.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
ssh-import-id==5.5
tzlocal==2.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
wtforms-validators==1.0.0
